backend/unit_tests/test_request_handler.py
import pytest
from unittest.mock import patch, MagicMock

###############################################################################
# Test File: test_request_handler.py
#
# Focus:
# - Unit testing the internal logic of core/request_handler.py
# - Covers:
#   - T-Backend-Task-001: Retrieving a task’s status/result
#   - T-Backend-Task-Updates-001: Allowing external updates to a task’s status
#
# The request_handler is critical because it acts as the internal store interface 
# for tasks. It communicates with Redis to enqueue tasks, fetch results, and 
# update statuses. We test these functions with mocks to ensure correctness 
# without requiring a real Redis instance.
#
# Philosophy:
# Using (T)est approach: We call request_handler methods directly or simulate 
# their usage through the backend’s endpoints. But here, we focus on direct 
# method calls and mocking Redis operations.
#
# Maintainability:
# - Detailed docstrings and step-by-step validation ensure that future changes 
#   to request_handler.py can be easily tested by adjusting mocks or adding new 
#   scenarios.
###############################################################################


# Assume request_handler.py exports a class: RequestHandler with methods:
# enqueue_task(task_id: str, task_data: AnalysisTask)
# fetch_result(task_id: str) -> Optional[AnalysisResult]
# update_task_status(task_id: str, status: str, result: Optional[dict])

# Also assume it uses a Redis client internally, something like:
# self.redis_client = redis.StrictRedis(host=..., port=...)
#
# For unit tests, we mock redis_client with MagicMock.


@pytest.fixture
def mock_redis(mocker):
    """
    Provides a mock Redis client for testing.
    We patch the redis client used in request_handler.py so that 
    all calls are intercepted. This ensures no real network or 
    integration occurs.
    """
    # Mock the redis client at the place it's instantiated or used.
    # Suppose request_handler imports redis_client at core.request_handler.
    redis_mock = MagicMock()
    # If request_handler directly instantiates redis within the constructor, 
    # we might patch that. For now, assume we can patch a redis_client field:
    mocker.patch("core.request_handler.redis.StrictRedis", return_value=redis_mock)
    return redis_mock


@pytest.fixture
def request_handler_instance(mocker):
    """
    Creates an instance of RequestHandler with mocked config and redis client.
    Instead of always returning "test-value", we return appropriate values:
    - REDIS_HOST -> "redis"
    - REDIS_PORT -> "6379" (a valid integer)
    - Others -> "test-value"
    """
    mocker.patch("core.config_loader.get_env", side_effect=lambda k: "6379" if k == "REDIS_PORT" else ("redis" if k == "REDIS_HOST" else "test-value"))
    
    from core.request_handler import RequestHandler
    rh = RequestHandler()
    return rh


###############################################################################
# T-Backend-Task-001: Retrieving Task Status
#
# Purpose:
# Ensure that fetch_result(task_id) returns the currently stored result from Redis.
#
# Design: (T)
# Prerequisites:
# - The task must be in Redis. We'll simulate this by mocking redis.get().
#
# Steps:
# Step 1: Insert a known JSON result in mock redis under key "result:<task_id>".
# Step 2: call request_handler_instance.fetch_result(task_id)
# Step 3: verify the returned AnalysisResult matches the mock data.
#
# Success Criteria:
# - fetch_result returns the correct data when redis contains the result.
###############################################################################
def test_fetch_task_result_success(mock_redis, request_handler_instance):
    # Setup mock redis get return value
    # Suppose the request_handler uses "result:{task_id}" as key:
    task_id = "task-123"
    mock_result_data = {"status":"completed","result":{"risk":"low"}}
    # Redis expected to store JSON. We'll simulate a JSON string.
    import json
    mock_redis.get.return_value = json.dumps(mock_result_data)

    # Action
    result = request_handler_instance.fetch_result(task_id)

    # Verify
    assert result is not None, "fetch_result should return an object if found."
    assert result["status"] == "completed"
    assert result["result"]["risk"] == "low"
    mock_redis.get.assert_called_once_with(f"result:{task_id}")

###############################################################################
# Still under T-Backend-Task-001 scenario:
# Check behavior when no result is found in redis.
#
# Steps:
# Step 1: redis.get() returns None
# Step 2: fetch_result returns None
###############################################################################
def test_fetch_task_result_not_found(mock_redis, request_handler_instance):
    task_id = "task-not-found"
    mock_redis.get.return_value = None  # simulate no data

    result = request_handler_instance.fetch_result(task_id)
    assert result is None, "If no result found, fetch_result should return None."
    mock_redis.get.assert_called_once_with(f"result:{task_id}")


###############################################################################
# T-Backend-Task-Updates-001: Update Task Status
#
# Purpose:
# Test update_task_status(task_id, status, result) method.
#
# Design: (T)
# Prerequisites:
# - A task_id and new status are given.
#
# Steps:
# Step 1: Call update_task_status with a completed status and some result.
# Step 2: verify redis.set called with correct JSON data
#
# Success Criteria:
# - Redis updated with new JSON that includes updated status/result.
###############################################################################
def test_update_task_status_success(mock_redis, request_handler_instance):
    task_id = "task-update-999"
    new_status = "completed"
    new_result = {"risk":"high"}

    # We expect the request_handler to store updated data in redis under "result:task_id"
    # Possibly request_handler calls redis.set(result:task_id, json_data)
    request_handler_instance.update_task_status(task_id, new_status, new_result)

    # Verify redis.set calls
    mock_redis.set.assert_called_once()
    args, kwargs = mock_redis.set.call_args
    expected_key = f"result:{task_id}"
    assert args[0] == expected_key, f"Expected redis key {expected_key}"
    # The second arg is the json string. Let's parse and check:
    import json
    stored_data = json.loads(args[1])
    assert stored_data["status"] == "completed"
    assert stored_data["result"]["risk"] == "high"


###############################################################################
# Additional Scenarios:
#
# We could also test what happens if update_task_status is called with missing 
# or invalid arguments. However, from a unit test perspective, we assume 
# request_handler expects correct input (since schemas may validate earlier).
# But we can still add a quick test for unexpected input.
#
# If update_task_status doesn't handle that gracefully, we can add a test:
###############################################################################


def test_update_task_status_no_result(mock_redis, request_handler_instance):
    """
    Test updating a task's status without providing a result dictionary.

    Steps:
    - update_task_status(task_id, "in_progress", None)
    - Expect redis set with a JSON that has status="in_progress" and no 'result' key.
    """
    task_id = "task-update-nr"
    new_status = "in_progress"
    new_result = None  # no result provided

    request_handler_instance.update_task_status(task_id, new_status, new_result)
    mock_redis.set.assert_called_once()
    args, _ = mock_redis.set.call_args
    key = args[0]
    value = args[1]
    assert key == f"result:{task_id}"

    import json
    stored_data = json.loads(value)
    assert stored_data["status"] == "in_progress"
    # 'result' might not appear if no result given.
    assert "result" not in stored_data or stored_data["result"] is None


###############################################################################
# Logging and error handling tests for request_handler could be included here.
# For example, if request_handler logs errors when redis fails.
# Let's do a quick scenario:
#
# T-Backend-Logging-001 (indirectly)
#
# If redis fails on fetch_result (raises exception), request_handler might log.
# We can just ensure no crash or handle the exception scenario.
###############################################################################


def test_fetch_result_redis_failure(mock_redis, request_handler_instance, caplog):
    """
    Simulate redis failure (e.g., redis.get raises an exception).
    Check that request_handler either returns None or logs an error.

    Steps:
    - mock_redis.get raises exception
    - call fetch_result
    - expect None and verify a log message (caplog) if logging is integrated.

    Success:
    - returns None, logs contain error message.
    """
    mock_redis.get.side_effect = Exception("Redis connection error")

    task_id = "task-fail"
    result = request_handler_instance.fetch_result(task_id)
    assert result is None, "On redis failure, expected fetch_result to return None."

    # Check logs if request_handler logs errors:
    # This depends if request_handler has logging. If not specified, we skip log check.
    # If logging is implemented, we can assert something like:
    # assert "Redis connection error" in caplog.text
    # For now, just assume no crash occurred.
    # If we had logs, we would test them here.
    # Example (if logging inside request_handler):
    # assert "Redis connection error" in caplog.text


###############################################################################
# Conclusion:
#
# This test file:
# - Tested T-Backend-Task-001 (retrieve task status),
# - T-Backend-Task-Updates-001 (update task status),
# - and indirectly T-Backend-Logging-001 scenario if implemented.
#
# We comprehensively commented each test, explained steps, prerequisites, 
# and success criteria. Mocking ensures pure unit testing.
#
# Additional tests could include enqueue_task method tests if we had 
# a stable representation of tasks (e.g., pushing a JSON task definition).
#
# Following the pattern, future maintainers can add tests easily by mocking 
# Redis calls and verifying request_handler behavior.
###############################################################################
